requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
mysql-connector-python==8.1.0
pip==21.1.2
requests==2.28.2
setuptools=57.0.0
soupsieve=2.4.1
urllib3==1.26.15
wheel=0.36.2
